Refactored GamePlaying
